Enable animations to halt without destroy promises
Previously, calling reset on an animation destroyed all of the promises
on the animation as well as the tick generator.  This isn't right - we
want to hang on to the promises so that we can stop and resume the
animation.  Add another method 'halt' so that we can reset the animation
completely if we so desire.